move tracks to view model
